fix(S3FIFO): keep the 64-bit frequency when the main FIFO evicts
S3FIFO, S3FIFOv0 and Clock2QPlus count an object's reads in an int64_t,
and every hit increments it without bound until the main FIFO next
passes the object. The main-FIFO eviction loop copies that count into
an int before testing freq >= 1. An object read 2^31 times since its
last pass narrows to a negative value, so the loop evicts the hottest
object in the cache instead of reinserting it.

The loops now copy the count into an int64_t. The new test gives the
oldest main-FIFO object 2^31 reads and checks that the next eviction
reinserts it and evicts the object behind it.

+++ libCacheSim/cache/eviction/Clock2QPlus.c
@@ -367,7 +367,7 @@ static void Clock2QPlus_evict_main(cache_t *cache, const request_t *req) {
   while (!has_evicted && main->get_occupied_byte(main) > 0) {
     cache_obj_t *obj_to_evict = main->to_evict(main, req);
     DEBUG_ASSERT(obj_to_evict != NULL);
-    int freq = obj_to_evict->Clock2QPlus.freq;
+    int64_t freq = obj_to_evict->Clock2QPlus.freq;
     copy_cache_obj_to_request(params->req_local, obj_to_evict);
     if (freq >= 1) {
       // we need to evict first because the object to insert has the same obj_id
